Basic Movement and Swapping Functionality
Imported Mannequin as a mesh placeholder. Basic character and camera movement has been implemented, and characters can be swapped between utilizing the PlayerControllerBP and Level Blueprint. There are some issues, such as continued animation on dispossession and that a reference to each possessable character must be manually added to the Level Blueprint, but it's a start.