Statuses and Status Pools Work Now
Pools are a single class with a public variable that defines what status it inflicts.
Statuses are made of a hierarchy, each falling under a category of Continuous effect or Interval effect. So far, Regen, Poison, and Slow have been implemented and appear to work. Additionally, they function differently depending on whether they came from a Pool or not (having not come from a pool hasn't been tested yet).