New PlayerSpawner
Just add in a PartySpawner and it will communicate with the GameMode to spawn an appropriate party. The only problem with this is that you can't customize the spawn locations with this method, as they are generated at runtime. This could be fixed, though it would then disallow party-size being determined at runtime (tiny issue and not important for our scope now, but it would be nice to me if it could handle that)