Basic Player Indicator
Added a basic visual indicator to the playable characters.
Currently, it's a circle that rotates around the base of the character. There is functionality to have a trail, but there are some minor issues with it, so it's been disabled for now. To reenable it to see what it looks like, go to the CharacterHighlighterBP and disconnect the deactivation node.