Pool Effects: No More Wireframes!
Status effects have begun to be added to the pools. The framework should allow for easy burn and slow effects.
There is a general boundary indicator. It looks kind of ugly now, but it does its job in conveying a sense of where the status effect is.
Effects exist for regen and poison pools!
Also, status effects each have enumerations now. The initial reason for this was so that a pool doesn't have to use an expanding series of casts and branches to determine what particle system to use. Instead, a switch can be used.