Fixed Player Icons
Okay pretty sure they work now. Solution doesn't feel great, but it works. I have to spawn instances of the objects in the world (I put them way below in the -Z axis) for them to display, and it doesn't like it if I destroy them or spawn them on top of each other, so each of them spawn lower than the previous. Not sure how this would work if other widgets also need to do this with other icons